requirements.txt
django>=2.2.9,<2.3.0
djangorestframework>=3.11.1,<3.12.0
psycopg2>=2.8.5,<2.9.0
Pillow>=5.3.0,<5.4.0

